# wiki/assets/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ops Agent API",
    "version": "0.1.0",
    "description": "Generated from registered Ops Agent routes. Middleware authentication, manually parsed request models and SSE media types are documented. WebSocket frames are described separately. Documentation viewer does not execute requests."
  },
  "paths": {
...
    "/api/console/run": {
      "post": {
        "summary": "Run Console Agent",
        "operationId": "run_console_agent_api_console_run_post",
        "responses": {
          "200": {
            "description": "Server-sent events; see the streaming guide.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "console"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsoleRunRequest"
              }
            }
          }
        }
      }
    },
    "/api/console/conversations/{conversation_id}/runtimes": {
      "get": {
        "summary": "List Conversation Runtimes",
        "operationId": "list_conversation_runtimes_api_console_conversations__conversation_id__runtimes_get",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuntimeSummaryView"
                  },
                  "title": "Response List Conversation Runtimes Api Console Conversations  Conversation Id  Runtimes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "console"
        ]
      }
    },
    "/api/console/runtimes/{runtime_id}/snapshot": {
      "get": {
        "summary": "Get Runtime Snapshot",
        "operationId": "get_runtime_snapshot_api_console_runtimes__runtime_id__snapshot_get",
        "parameters": [
          {
            "name": "runtime_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Runtime Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeSnapshotView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "console"
        ]
      }
    },
    "/api/console/runtimes/{runtime_id}/events": {
      "get": {
        "summary": "Get Runtime Events",
        "operationId": "get_runtime_events_api_console_runtimes__runtime_id__events_get",
        "parameters": [
          {
            "name": "runtime_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Runtime Id"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Since"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeEventsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "console"
        ]
      }
    },
    "/api/console/runtimes/{runtime_id}/reconnect": {
      "post": {
        "summary": "Reconnect Runtime Stream",
        "operationId": "reconnect_runtime_stream_api_console_runtimes__runtime_id__reconnect_post",
        "parameters": [
          {
            "name": "runtime_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Runtime Id"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Since"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server-sent events; see the streaming guide.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "console"
        ]
      }
    },
    "/api/console/runtimes/{runtime_id}/cancel": {
      "post": {
        "summary": "Cancel Runtime",
        "operationId": "cancel_runtime_api_console_runtimes__runtime_id__cancel_post",
        "parameters": [
          {
            "name": "runtime_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Runtime Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Cancel Runtime Api Console Runtimes  Runtime Id  Cancel Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "console"
        ]
      }
    },
    "/api/console/runtimes/{runtime_id}/messages": {
      "post": {
        "summary": "Send Runtime Message",
        "operationId": "send_runtime_message_api_console_runtimes__runtime_id__messages_post",
        "parameters": [
          {
            "name": "runtime_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Runtime Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server-sent events; see the streaming guide.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "console"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeMessageRequest"
              }
            }
          }
        }
      }
    },
    "/api/console/terminal-requests/{request_id}/decision": {
      "post": {
        "summary": "Decide Terminal Request",
        "operationId": "decide_terminal_request_api_console_terminal_requests__request_id__decision_post",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server-sent events; see the streaming guide.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "console"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TerminalRequestDecisionRequest"
              }
            }
          }
        }
      }
    },
    "/api/console/approval": {
      "post": {
        "summary": "Approve Console Agent",
        "operationId": "approve_console_agent_api_console_approval_post",
        "responses": {
          "200": {
            "description": "Server-sent events; see the streaming guide.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "console"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsoleApprovalRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RuntimeEventsResponse": {
        "properties": {
          "latest_sequence": {
            "type": "integer",
            "title": "Latest Sequence"
          },
          "events": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "latest_sequence"
        ],
        "title": "RuntimeEventsResponse"
      },
...
      "RuntimeSummaryView": {
        "properties": {
          "runtime_id": {
            "type": "string",
            "title": "Runtime Id"
          },
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          },
          "asset_id": {
            "type": "integer",
            "title": "Asset Id"
          },
          "terminal_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terminal Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "run_state": {
            "type": "string",
            "title": "Run State",
            "default": "queued"
          },
          "loaded_skill_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Loaded Skill Name"
          },
          "current_step_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Step Id"
          },
          "pending_approval_step_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Approval Step Id"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "runtime_id",
          "conversation_id",
          "asset_id",
          "status",
          "updated_at"
        ],
        "title": "RuntimeSummaryView"
      },
...
      "ConsoleRunRequest": {
        "properties": {
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "mode": {
            "default": "standard",
            "enum": [
              "standard",
              "incident"
            ],
            "title": "Mode",
            "type": "string"
          },
          "currentEvents": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Currentevents",
            "type": "array"
          },
          "asset_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Asset Id"
          },
          "terminal_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Terminal Id"
          },
          "conversation_id": {
            "default": "console",
            "title": "Conversation Id",
            "type": "string"
          },
          "user_event_id": {
            "anyOf": [
              {
                "pattern": "^user-[A-Za-z0-9_-]{8,100}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "User Event Id"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Model Name"
          },
          "selected_skill_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Selected Skill Name"
          },
          "terminal_context": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Terminal Context"
          }
        },
        "required": [
          "prompt"
        ],
        "title": "ConsoleRunRequest",
        "type": "object"
      },
      "RuntimeMessageRequest": {
        "properties": {
          "message": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "RuntimeMessageRequest",
        "type": "object"
      },
      "TerminalRequestDecisionRequest": {
        "properties": {
          "runtimeId": {
            "title": "Runtimeid",
            "type": "string"
          },
          "approvalToken": {
            "title": "Approvaltoken",
            "type": "string"
          },
          "approved": {
            "title": "Approved",
            "type": "boolean"
          }
        },
        "required": [
          "runtimeId",
          "approvalToken",
          "approved"
        ],
        "title": "TerminalRequestDecisionRequest",
        "type": "object"
      },
      "ConsoleApprovalRequest": {
        "properties": {
          "runtime_id": {
            "title": "Runtime Id",
            "type": "string"
          },
          "approved": {
            "title": "Approved",
            "type": "boolean"
          },
          "approval_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Approval Token"
          },
          "allow_prefix": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Allow Prefix"
          },
          "guidance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Guidance"
          }
        },
        "required": [
          "runtime_id",
          "approved"
        ],
        "title": "ConsoleApprovalRequest",
        "type": "object"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}